+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Changelog
 
-
-## Unreleased (0.51.0)
+## Unreleased (0.52.0)
 
 ### 🚨 Breaking changes
 - [](https://github.com/vegaprotocol/protos/issues/) -
@@ -10,13 +9,24 @@
 - [](https://github.com/vegaprotocol/protos/pull/) -
 
 ### 🛠 Improvements
+- [418](https://github.com/vegaprotocol/protos/issues/418) - Add support for cursor based pagination in `Data Node API V2`
+- [](https://github.com/vegaprotocol/protos/pull/) -
+
+### 🐛 Fixes
+- [](https://github.com/vegaprotocol/protos/pull/) -
+
+
+## 0.51.0
+
+### 🛠 Improvements
 - [406](https://github.com/vegaprotocol/protos/issues/406) - Add content for asset bundle query end point
 - [413](https://github.com/vegaprotocol/protos/issues/413) - Add support for sending transfers to asset/market
 - [401](https://github.com/vegaprotocol/protos/pull/401) - Add rationale on ProposalSubmission
 - [5423](https://github.com/vegaprotocol/vega/issue/5423) - Add new fields for equity shares to `Snapshot` types.
-- [418](https://github.com/vegaprotocol/protos/issues/418) - Add support for cursor based pagination in `Data Node API V2`
 - [420](https://github.com/vegaprotocol/protos/issues/420) - Enable recurring transfer to automatically dispatch reward pro-rata based on market metric
-- 
+- [424](https://github.com/vegaprotocol/protos/issues/424) - Add v1 and v2 endpoints to list all oracle data
+- [424](https://github.com/vegaprotocol/protos/issues/424) - Add missing v2 endpoints for oracle specs and data
+- [424](https://github.com/vegaprotocol/protos/issues/424) - Add support for pagination on v1 endpoints for oracle specs and data
 
 ### 🐛 Fixes
 - [](https://github.com/vegaprotocol/protos/pull/) -